refactor: update namespace from 'Grids' to 'Grid' for consistency
